update server.go +error.html |add errors.go + handler.go
REFACTORING DU CODE
