--- handlers/connection.go
@@ -1,21 +1,0 @@
-package handlers
-
-import (
-	"log"
-	"net/http"
-	"text/template"
-
-	_ "github.com/mattn/go-sqlite3"
-)
-
-func ConnexionPage(w http.ResponseWriter, r *http.Request) {
-	// Déclaration des fichiers à parser
-	t, err := template.ParseFiles("templates/connection.html", "./templates/layouts/sidebar.html", "./templates/layouts/header.html")
-	if err != nil {
-		Color(3, "[SERVER_INFO_PAGE] : 🟠 Template execution : ")
-		log.Fatalf("%s", err)
-		return
-	}
-	Color(1, "[SERVER_INFO_PAGE] : 🟢 Page 'connexion'")
-	t.Execute(w, nil)
-}
