--- handlers/inscription.go
@@ -1,20 +1,0 @@
-package handlers
-
-import (
-	"log"
-	"net/http"
-	"text/template"
-)
-
-func InscriptionPage(w http.ResponseWriter, r *http.Request) {
-
-	// Déclaration des fichiers à parser
-	t, err := template.ParseFiles("templates/inscription.html", "./templates/layouts/sidebar.html", "./templates/layouts/header.html")
-	if err != nil {
-		Color(3, "[SERVER_INFO_PAGE] : 🟠 Template execution : ")
-		log.Fatalf("%s", err)
-		return
-	}
-	Color(1, "[SERVER_INFO_PAGE] : 🟢 Page 'inscription'")
-	t.Execute(w, nil)
-}
