+++ templates/categories.html
@@ -6,7 +6,7 @@
         <meta name="viewport" content="width=device-width, initial-scale=1.0">
         <link href='https://unpkg.com/boxicons@2.0.7/css/boxicons.min.css' rel='stylesheet'>
         <link rel="stylesheet" href="../static/css/all_categories.css">
-        <title>All_categories</title>
+        <title>Categories</title>
     </head>
     <body>
         {{template "sidebar"}} {{template "header"}}
