+++ templates/layouts/error404.html
@@ -7,6 +7,6 @@
     <title>Document</title>
 </head>
 <body>
-    <h1>error 404</h1>
+    <h1 class="test">Error 404 : Cette page n'existe pas.</h1>
 </body>
 </html>
