+++ templates/layouts/error500.html
@@ -7,6 +7,6 @@
     <title>Document</title>
 </head>
 <body>
-    <h1>error 500</h1>
+    <h1>Error 500 : Erreur du serveur.</h1>
 </body>
 </html>
